Use musl to provide ARM release binaries (#32)
It was a pain, but we can now generate ARM64 binaries for the linux server by relying on musl.
It is a global rework of the build and goreleaser scripts. We are forced to split .goreleaser-linux.yml in two distinct files to be able to specify the path to ld. I did not find any other way to set it through flags that works its way through go build.