Support proxy jump (#94)
This adds support for classical, secure, UDP-based Proxy-Jump.
Let's say we want to connect from A to C through B, B being the proxy. This works with the client located on A connecting to B and establishing UDP forwarding from towards C. By doing so, the client can send regular encrypted QUIC packets through the forwarded socket to connect to C. B cannot manipulate the QUIC packets are they are encrypted and authenticated.

The drawbacks of the approach are:

- quic-go datagram are slow
- we tunnel a congestion-controlled protocol through another congestion-controlled protocol which might cause problems, especially that quic-go is quite prompt to drop UDP datagrams.

So this will work the best for interactive sessions. If troughput is needed, we may want to look at HTTP reverse proxies instead.

Fixes #44

* add -proxy-jump switch

* add proxy-jump integration tests

* parse the UDPProxyJump ssh config option that works like the SSHv2 ProxyJump, without variables (%h, %p, etc are not supported). If used within a classical OpenSSH config file, add `IgnoreUnknown UDPProxyJump` in the config file to avoid OpenSSH errors with this new option.

* bump quic-go version for quic-in-quic capabilities

* bump go version to match what is supported by quic-go

* bump setup-go version for CI build scripts