README: remove ssh3 name decomposition
I removed that to address people saying that SSH + H3 could be something else than SSH3 (e.g. SSHTTP3, SSHH3, or other weird names).

The name really was chosen because it first perfectly the fact that it can be a new version of the protocol *and* it uses H3 underneath. But in the same time I would like to spare my energy, so just let's remove the sentence.